cursor-movement-using-eyes/requirements.txt
# Voice agent
speechrecognition
pyaudio
groq
python-dotenv
elevenlabs
pygame

# Eye tracking (from Shift)
opencv-python
mediapipe
pyautogui
numpy
sounddevice
pyobjc-framework-Vision
